+++ backend/lost/pyapi/pipe_elements.py
@@ -80,6 +80,19 @@ class AnnoTask(Element):
         lbl_df = pd.concat(lbl_list)
         return lbl_df
 
+    @property
+    def lbl_map(self):
+        '''dict: Map lbl_name to idx
+
+        Note:
+            All label names will be mapped to lower case!
+        '''
+        df = self.possible_label_df
+        my_map = dict()
+        def create_map(x):
+            my_map[x['name'].lower()] = x['idx'] 
+        df.apply(lambda x: create_map(x),axis=1)
+        return my_map
     
     @property
     def instructions(self):